Upgrade collection id from int to string as per recent Unsplash API changes

+++ app/src/main/java/com/b_lam/resplash/ui/collection/detail/CollectionDetailActivity.kt
@@ -42,7 +42,7 @@ class CollectionDetailActivity : BaseActivity(R.layout.activity_collection_detai
 
         when {
             collection != null -> viewModel.setCollection(collection)
-            collectionId != null -> viewModel.getCollection(collectionId.toInt())
+            collectionId != null -> viewModel.getCollection(collectionId)
             else -> finish()
         }
